+++ sdk/communication/azure-communication-chat/src/main/java/com/azure/android/communication/chat/models/TypingIndicatorReceivedEvent.java
@@ -26,6 +26,12 @@ public final class TypingIndicatorReceivedEvent extends ChatUserEvent {
     @JsonProperty(value = "originalArrivalTime")
     private OffsetDateTime receivedOn;
 
+    /**
+     * The display name of the event sender.
+     */
+    @JsonProperty(value = "senderDisplayName")
+    private String senderDisplayName;
+
     static {
         EventAccessorHelper.setTypingIndicatorReceivedEventAccessor(event -> {
             TypingIndicatorReceivedEvent typingIndicatorReceivedEvent = (TypingIndicatorReceivedEvent) event;
@@ -54,5 +60,14 @@ public final class TypingIndicatorReceivedEvent extends ChatUserEvent {
      */
     public OffsetDateTime getReceivedOn() {
         return receivedOn;
+    }
+
+    /**
+     * Gets the display name of the event sender.
+     *
+     * @return Value of The display name of the event sender.
+     */
+    public String getSenderDisplayName() {
+        return senderDisplayName;
     }
 }